fixed leaks

+++ include/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: elenavoronin <elnvoronin@gmail.com>          +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/02/16 15:51:41 by yitoh         #+#    #+#                 */
-/*   Updated: 2024/03/19 16:11:25 by evoronin      ########   odam.nl         */
+/*   Updated: 2024/03/20 14:03:11 by evoronin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,8 +32,8 @@ typedef struct s_map
 	int			**map2d;
 	int			*floor;
 	int			*ceiling;
-	size_t		depth;
-	size_t		width;
+	int			depth;
+	int			width;
 	int			px; //player's position
 	int			py;
 }	t_map;
